Allow PassFF Integration
PassFF is a browser plugin that uses the password-store.org based
password database to fill forms on firefox. For this it guesses from the
form elements which fields should be entered. Apparently the URI-field
names are filled with usernames or the url of the site that is visisted.
As such the passFF integration breaks usability.

This renames the tr_uri to tr_redmine to avoid this behavior. Also the
label is adjusted for clarification